dev(drivers): refactor drivers
