+++ src/drivers/tty/internal/tty.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <drivers/tty.h>
+
+// ============================================================================
+// DEFINE AND MACRO
+// ============================================================================
+
+// ============================================================================
+// STRUCT
+// ============================================================================
+
+// ============================================================================
+// VARIABLES GLOBALES
+// ============================================================================
+
+// ============================================================================
+// INTERNAL APIs
+// ============================================================================
+
+static bool ft_strequ(const char *s1, const char *s2);
+static void print_help(void);
