+++ src/drivers/vga/vga.c
@@ -1,9 +1,4 @@
-#include <libft.h>
-
-#include "io.h"
-#include "printk.h"
-#include "tty.h"
-#include "vga.h"
+#include "internal/vga.h"
 
 #define KERNEL_BANNER                                                                              \
 	"  /$$   /$$   /$$   /$$               /$$   /$$\n"                                            \
@@ -18,6 +13,88 @@
 struct s_cursor g_cursor;
 static uint8_t  vga_mode = VGA_DEFAULT_MODE;
 
+static int ft_putchar(char c)
+{
+	switch (c) {
+	case '\n':
+		g_cursor.y++;
+		fallthrough;
+
+	case '\r':
+		g_cursor.x = 0;
+		break;
+
+	case '\t':
+		do {
+			ft_putchar(' ');
+		} while (g_cursor.x % 4);
+		break;
+
+	default:
+		vga_set_char(g_cursor.x++, g_cursor.y, c);
+	}
+
+	while (g_cursor.x >= VGA_WIDTH) {
+		g_cursor.x -= VGA_WIDTH;
+		g_cursor.y++;
+	}
+
+	while (g_cursor.y >= VGA_HEIGHT) {
+		vga_scroll_down();
+		g_cursor.y--;
+	}
+
+	vga_set_cursor_position(g_cursor.x, g_cursor.y);
+	return 1;
+}
+
+static int ft_putstr(char *str)
+{
+	int i = -1;
+
+	if (str == NULL)
+		return (ft_putstr("(null)"));
+	while (str[++i])
+		ft_putchar(str[i]);
+	return (i);
+}
+
+static int ft_putnbr(int nb)
+{
+	if (nb == -2147483648)
+		return (ft_putnbr(-21474) + ft_putnbr(83648));
+	if (nb < 0)
+		return (ft_putchar('-') + ft_putnbr(nb * -1));
+	else if (nb >= 10)
+		return (ft_putnbr(nb / 10) + ft_putnbr(nb % 10));
+	else
+		return (ft_putchar(nb + '0'));
+}
+
+static int ft_putunsignednbr(uint32_t nb)
+{
+	if (nb >= 10)
+		return (ft_putnbr(nb / 10) + ft_putnbr(nb % 10));
+	else
+		return (ft_putchar(nb + '0'));
+}
+
+static int ft_puthexa(unsigned long nb, bool upper, bool addr)
+{
+	char *dict = upper ? "0123456789ABCDEF" : "0123456789abcdef";
+
+	if (addr)
+		return (ft_putstr("0x") + ft_puthexa(nb, upper, 0));
+	else if (nb < 16)
+		return (ft_putchar(dict[nb]));
+	else
+		return (ft_puthexa(nb / 16, upper, 0) + ft_puthexa(nb % 16, upper, 0));
+}
+
+// Internal APIs
+//////////////////////////////////////////
+// External APIs
+
 uint8_t vga_get_mode(void) { return vga_mode; }
 
 void vga_setup_default_screen()
@@ -26,7 +103,7 @@ void vga_setup_default_screen()
 	vga_enable_cursor(14, 15);
 	vga_set_cursor_position(0, 1);
 
-	printk(KERNEL_BANNER "\n" TTY_PROMPT);
+	vga_printf(KERNEL_BANNER "\n" TTY_PROMPT);
 }
 
 void vga_enable_cursor(uint8_t cursor_start, uint8_t cursor_end)
@@ -83,4 +160,49 @@ void vga_clear(void)
 	for (int y = 0; y < VGA_HEIGHT; y++)
 		for (int x = 0; x < VGA_WIDTH; x++)
 			*VGA_ENTRY(x, y) = (vga_entry){.character = 0x00, .mode = vga_mode};
+}
+
+void vga_printf(const char *fmt, ...)
+{
+	va_list ap;
+	va_start(ap, fmt);
+
+	for (unsigned long i = 0; fmt[i]; i++) {
+		if (fmt[i] == '%') {
+			switch (fmt[++i]) {
+			case 's':
+				ft_putstr(va_arg(ap, char *));
+				break;
+
+			case 'i':
+			case 'd':
+				ft_putnbr(va_arg(ap, int));
+				break;
+
+			case 'u':
+				ft_putunsignednbr(va_arg(ap, uint32_t));
+				break;
+
+			case 'x':
+				ft_puthexa(va_arg(ap, uint32_t), false, false);
+				break;
+
+			case 'X':
+				ft_puthexa(va_arg(ap, uint32_t), true, false);
+				break;
+
+			case 'p':
+				ft_puthexa(va_arg(ap, uint32_t), false, true);
+				break;
+
+			case 'c':
+				ft_putchar(va_arg(ap, uint32_t));
+				break;
+
+			default:
+				break;
+			}
+		} else
+			ft_putchar(fmt[i]);
+	}
 }
